% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_inflation.R
\name{adjust_inflation}
\alias{adjust_inflation}
\title{Adjust species value for inflation}
\usage{
adjust_inflation(comland, refyear, refmonth)
}
\arguments{
\item{comland}{Data frame. master data frame containing species landings}

\item{refyear}{Integer. Reference year}

\item{refmonth}{Integer. Reference month}
}
\value{
comland data frame adjusted for inflation
}
\description{
Reads in data from Bureau of Labor statistics website and adjusts species value
}
\details{
What data is fetched?
}
\keyword{internal}
