% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_epu.R
\name{aggregate_by_epu}
\alias{aggregate_by_epu}
\title{aggregate data by epu}
\usage{
aggregate_by_epu(comland.agg, nafoland.agg, EPUS, foreign)
}
\arguments{
\item{comland.agg}{Data frame. master data frame containing species landings}

\item{nafoland.agg}{Data frame. processed NAFO data}

\item{EPUS}{List. Designates the stat areas that comprise an EPU. Default = EPUs (lazily loaded data)}

\item{foreign}{Character String. Mark foreign landings and keep seperate. Default = "y"}
}
\value{
Aggregated comland data
}
\description{
aggregate by epu
}
\keyword{internal}
