% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gear.R
\name{aggregate_gear}
\alias{aggregate_gear}
\title{Assign landing records to an aggregated gear fleet}
\usage{
aggregate_gear(comData, userGears, fleetDescription)
}
\arguments{
\item{userGears}{Data frame. Definitions to aggregate NEGEAR2 to user defined
gears}

\item{fleetDescription}{Character. Name of column in userGears that defines 
the new gears.}

\item{comdata}{Data set generated by \code{get_comland_data} or 
\code{get_comdisc_data}}
}
\description{
Takes the output from \code{get_comland_data} or \code{get_comdisc_data} and 
further aggregates from NEGEAR2 codes to a user defined fleet.
}
