% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_area.R
\name{assign_area}
\alias{assign_area}
\title{Assigns points to polygon}
\usage{
assign_area(comdiscData, areaPolygon, areaDescription, na.keep = F)
}
\arguments{
\item{na.keep}{Boolean. Logical value to indicate whether original strata names
should be retained.}
}
\value{
Returns a \code{comdiscData} data.table with one additional column labeled
 with the value of \code{areaDescription}

\item{areaDescription}{The name of the region (found in \code{areaPolygon})
 that a record in \code{surveyData} is assigned to}
}
\description{
Assign observer data (points, lat and lon) to designated regions (polygons) from a shape file.
}
\seealso{
Other comdisc: 
\code{\link{calc_DK}()}
}
\concept{comdisc}
