% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_catch_area.R
\name{assign_catch_area}
\alias{assign_catch_area}
\title{assign unknown catch using known catch characteristics}
\usage{
assign_catch_area(comland.agg)
}
\value{
updated comland
}
\description{
Uses area. Expand .....
}
