% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_catch_size.R
\name{assign_catch_size}
\alias{assign_catch_size}
\title{assign unknown catch using known catch characteristics}
\usage{
assign_catch_size(comland.agg)
}
\value{
updated comland
}
\description{
Uses size. Expand .....
}
