% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_unknown.R
\name{assign_unknown}
\alias{assign_unknown}
\title{Assigns unknown catch variables}
\usage{
assign_unknown(
  comData,
  unkVar,
  knStrata = c("HY", "QY", "MONTH", "NEGEAR", "TONCL1", "AREA")
)
}
\arguments{
\item{unkVar}{Vector. Catch variables to be imputed.  Should be in the order
you would like them to be solved.}

\item{knStrata}{Vector. Catch variables to be used to impute \code{unkVar}. An
\code{unkVar} can be included in this list if it is used to solve a second 
\code{unkVar}.  Should be in the order of most restrictive to least restrictive.}
}
\value{
Returns a \code{comlandData} data.table.
}
\description{
Impute unknown variables associated with landings.
}
\concept{comland}
