% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_discards.R
\name{calc_discards}
\alias{calc_discards}
\title{Calculate stratified mean}
\usage{
calc_discards(comland, dk, areaDescription, fleetDescription)
}
\arguments{
\item{tidy}{Boolean. Return output in long format (Default = F).}

\item{returnPrepData}{Boolean. Return both \code{stratmeanData} and \code{prepData}
as a list object. The default (F) returns only the \code{stratmeanData} as a
\code{data.table}.}
}
\description{
Calculates the stratified mean. Details of method found here ...
}
\examples{
\dontrun{
# Pull data and apply conversion corrections
data <- get_survdat_data(channel)
# Calculate stratified mean for specific survey strata for the SPRING season
calc_stratified_mean(surveyData=data$survdat, filterByArea=c(1220, 1240, 1260:1290,1360:1400),filterBySeason = "SPRING")

# Calculate stratified mean for area defined by EPU regions, for all seasons ("SPRING", "FALL")
# Read in EPU shapefile (loaded as part of the package)
area <- sf::st_read(dsn = system.file("extdata","EPU.shp",package="survdat"),quiet=T)
calc_stratified_mean(surveyData=data$survdat, areaPolygon=area, areaDescription="EPU", filterByArea="all",filterBySeason = "all")

}


}
\concept{survdat}
