% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comland.r
\name{comland}
\alias{comland}
\title{Comland.r Version now controlled by git - originally part of comcatch.r}
\usage{
comland(uid = "abeet", use.existing = "y",
  out.dir = here::here("output"))
}
\arguments{
\item{uid}{username for Oracle connection}

\item{use.existing}{String. Pull from database "n" or use existing pull "y"}

\item{out.dir}{path to directory where final output will be saved}
}
\description{
#Grab commercial landings data from US and Foreign countries (NAFO)
Need to fix menhaden data
SML
}
