% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_aggregate.R
\name{comland_aggregate}
\alias{comland_aggregate}
\title{aggregate data by time, gear, tonnage}
\usage{
comland_aggregate(comland, GEARS)
}
\arguments{
\item{comland}{Data frame. master data frame containing species landings}

\item{GEARS}{List. Designates the NEGEAR codes that comprise a fishing fleet.}
}
\value{
Aggreagted comland data
}
\description{
aggregate by quarter year, half year, major gear, and small/large TC
}
