% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_herring.R
\name{comland_herring}
\alias{comland_herring}
\title{Processes herring data}
\usage{
comland_herring(channel, comland)
}
\arguments{
\item{channel}{DBI object. connection object for database access}

\item{comland}{Data frame. master data frame containing species landings}
}
\value{
Processed Herring data added to comland
}
\description{
Herring Data comes from the state of Maine.
}
