% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_menhaden.R
\name{comland_menhaden}
\alias{comland_menhaden}
\title{Processes Menhaden data}
\usage{
comland_menhaden(comland)
}
\arguments{
\item{comland}{Data frame. master data frame containing species landings}
}
\value{
Processed Menhaden data
}
\description{
Fix menhaden records - data from Tom Miller/ Andre Bouchheister
}
