% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_nafo.R
\name{comland_nafo}
\alias{comland_nafo}
\title{Processes all NAFO data}
\usage{
comland_nafo(channel, skate.hake.nafo, GEARS)
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{skate.hake.nafo}{}

\item{GEARS}{List. Designates the NEGEAR codes that comprise a fishing fleet. Default = GEARs (lazily loaded data)}
}
\value{
Processed NAFO data
}
\description{
Dwonloads and reads in all NAFO data then aggreagtes it
}
