% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_separate_hakes.R
\name{comland_separate_hakes}
\alias{comland_separate_hakes}
\title{#Comcatch_skates_hakes.r}
\usage{
comland_separate_hakes(comland, skate.hake.us)
}
\arguments{
\item{comland}{Data frame. Master data frame containing species landings}

\item{skate.hake.us}{Data frame. Landings of skates and hakes in USA}
}
\value{
updated comland
}
\description{
Determine proportion of little/winter skates and silver hake in landings data 7/13
SML
}
