% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comland_skates_hakes.R
\name{comland_skates_hakes}
\alias{comland_skates_hakes}
\title{#Comcatch_skates_hakes.r}
\usage{
comland_skates_hakes()
}
\description{
Determine proportion of little/winter skates and silver hake in landings data 7/13
SML
}
