% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_unknowns.R
\name{comland_unknowns}
\alias{comland_unknowns}
\title{Processes unknown data}
\usage{
comland_unknowns(comland)
}
\arguments{
\item{comland}{Data frame. master data frame containing species landings}
}
\value{
Processed unknowns in comland
}
\description{
drop unknown species/landings
}
