% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_winter_little.R
\name{comland_winter_little}
\alias{comland_winter_little}
\title{#Comcatch_skates_hakes.r}
\usage{
comland_winter_little(comland, skate.hake.us)
}
\description{
Determine proportion of little/winter skates and silver hake in landings data 7/13
SML
}
