% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comlandr.R
\docType{package}
\name{comlandr-package}
\alias{comlandr}
\alias{comlandr-package}
\title{comlandr: Pull and process commercial fisheries data}
\description{
Suite of functions to pull and process landings and discard data in addition to
helper function to pull lookup tables
}
\details{
\itemize{
\item Pulls data from StockEff: by area, gear, species, year, month
\item pulls Herring data from state of Maine database
\item Defines fleet structure and assigns fleets to gear types
\item Define geographic areas and assigns statistical areas to geographic areas
\item Assigns landings to unknown gear, area, quarter/half year, size based on similar trips (based on M Palmer, ????)
\item Pulls and processes NAFO (Northwest Atlantic Fisheries Organization) data
\item Uses survey data to apportion hake complex into species
\item Uses survey data to apportion skate complex into species
\item Ajusts species value to specified date
}
}
\section{References}{


Palmer, M (????) Working doc reference and title

NAFO website: \url{https://www.nafo.int}

To learn more about using \code{comlandr}, visit \url{https://noaa-edab.github.io/comlandr/articles/comlandr.html} or click the index below
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NOAA-EDAB/comlandr}
  \item \url{https://noaa-edab.github.io/comlandr/}
  \item Report bugs at \url{https://github.com/NOAA-EDAB/comlandr/issues}
}

}
\author{
\strong{Maintainer}: Andy Beet \email{andrew.beet@noaa.gov} (\href{https://orcid.org/0000-0001-8270-7090}{ORCID})

Authors:
\itemize{
  \item Sean Lucey
}

}
\keyword{internal}
