% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate.R
\docType{data}
\name{deflate}
\alias{deflate}
\title{deflate:  Index of dollar inflation}
\format{
A Data frame
\describe{
\item{YEAR}{Year of index}
\item{MONTH}{Month of index}
\item{PPI}{Price Index}
}
}
\usage{
deflate
}
\description{
This needs to be pulled directly from gov website in future
This data set should be temporary since it will need to be updated each year
It isn't clear where from
}
\keyword{datasets}
