% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPUs.R
\docType{data}
\name{EPUs}
\alias{EPUs}
\title{EPUs:  Properties of Ecological Production Units (EPUs)}
\format{
A list of lists
\describe{
    \item{GB}{Georges Bank EPU}
    \item{GOM}{Gulf of Maine EPU}
    \item{MAB}{Mid Atlantic Bight EPU}
    \item{SS}{Scotian Shelf EPU}
}
}
\usage{
EPUs
}
\description{
Lists the four EPUs (Georges Bank (GB), Gulf of Maine (GOM), Mid Atlantic Bight (MAB), Scotian Shelf (SS)) with the statistical areas that make up each region and their area (km^2)
}
\section{List components}{


Each list item is a list itself containg the following elements:

\describe{
   \item{statAreas}{statistical areas that comprise the EPU}
   \item{area}{area in km^2 of toal EPU}
}
}

\keyword{datasets}
