% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEARs.R
\docType{data}
\name{GEARs}
\alias{GEARs}
\title{GEARs:  Gear type classifications}
\format{A list of NEGEAR codes assigned to "fleets"
\describe{
    \item{otter}{Otter Trawls: Bottom, Bottom paired, Bottom Twin, Hadock Separator}
    \item{dredge.sc}{Scallop Dredges}
    \item{pot}{Pots and Traps: Fish, eel, conch, octopus, hagfish , shrimp, lobster, crab}
    \item{longline}{Bottom and Pelagic Longlines}
    \item{seine}{Beach, Long, mackerel, menhaden, tuna, scottish Seines}
    \item{gillnet}{Fixed, Anchored, Sea bass, salmon, crab, floating, run around, other Gillnets}
    \item{midwater}{Midwater paired and unpaired otter trawls}
    \item{dredge.o}{Oyster, scrapes, Clam, Crab, conch, mussel, urchin, Quahog, hydraulic dredges}
}}
\usage{
GEARs
}
\description{
Lists the NESFC gear types (NEGEAR) attributed to a more general "fleet" characteristic
}
\keyword{datasets}
