% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comdisc_data.R
\name{get_comdisc_data}
\alias{get_comdisc_data}
\title{Extracts commercial discard data from Database}
\usage{
get_comdisc_data(
  channel,
  comland,
  aggArea = F,
  areaDescription = "EPU",
  propDescription = "MeanProp",
  aggGear = F,
  fleetDescription = "Fleet",
  extendTS = T
)
}
\arguments{
\item{channel}{an Object inherited from \code{\link[ROracle]{Oracle}}. This object is used to connect
to communicate with the database engine. (see \code{\link[dbutils]{connect_to_database}})}

\item{comland}{Data frame. Result of \code{get_comland_data}}

\item{aggArea}{boolean. Aggregate Statistical Areas into larger spatial units (Default = F)}

\item{areaDescription}{character string. Field name in \code{userAreas} denoting spatial unit. (Default = "EPU")}

\item{propDescription}{character string. Field name in \code{userAreas} denoting the scaling factor. (Default = "MeanProp")}

\item{aggGear}{boolean. Aggregate NEGEAR codes to larger "fleets" (Default = F)}

\item{fleetDescription}{character string. Field name in \code{userGears} denoting Fleet. (Default = "Fleet")}

\item{extendTS}{Boolean. Should the DK (Discard to kept) ratio be extended and applied
to landings beyond observer coverage time period (Discards started in 1989). Default = T}
}
\value{
Data frame (data.table) (n x 10)
Each row of the data.table represents a species record for a given tow/trip

\item{YEAR}{Year of trip/tow}
\item{MONTH}{Month of trip/tow}
\item{NEGEAR}{Fishing gear used on trip/tow}
\item{TONCL1}{Tonnage class of the fishing vessel}
\item{NESPP3}{Species code (3 charachters)}
\item{NESPP4}{Species code and market code (4 characters)}
\item{AREA}{Statistical area in which species was reportly caught}
\item{UTILCD}{Utilization code}
\item{SPPLIVLB}{live weight (landed = "n") or landed weight (landed="y") in lbs}
\item{SPPVALUE}{The value of landed catch to the nearest dollar (U.S.), paid to fisherman by dealer, for a given species.}
}
\description{
Connects to obdbs and pulls discard data, calculates discard to kept ratios,
and applies to landings data obtained using \code{get_comland_data}.
}
