% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comland_data.R
\name{get_comland_data}
\alias{get_comland_data}
\title{Extracts commercial data from Database}
\usage{
get_comland_data(
  channel,
  filterByYear = NA,
  filterByArea = NA,
  useLanded = T,
  removeParts = T,
  useHerringMaine = T,
  useForeign = T,
  refYear = NA,
  refMonth = NA,
  disagSkatesHakes = T,
  aggArea = F,
  userAreas = comlandr::mskeyAreas,
  areaDescription = "EPU",
  propDescription = "MeanProp",
  applyPropLand = T,
  applyPropValue = T,
  aggGear = F,
  userGears = comlandr::mskeyGears,
  fleetDescription = "Fleet",
  unkVar = c("MONTH", "NEGEAR", "AREA"),
  knStrata = c("HY", "QY", "MONTH", "NEGEAR", "TONCL2", "AREA")
)
}
\arguments{
\item{channel}{an Object inherited from \code{ROracle}. This object is used to connect
to communicate with the database engine. (see \code{\link[dbutils]{connect_to_database}})}

\item{filterByYear}{numeric vector}

\item{filterByArea}{numeric vector}

\item{useLanded}{boolean}

\item{removeParts}{boolean}

\item{useHerringMaine}{boolean}

\item{useForeign}{boolean}

\item{refYear}{numeric}

\item{refMonth}{numeric}

\item{aggArea}{boolean}

\item{userAreas}{data frame}

\item{areaDescription}{character string}

\item{propDescription}{character string}

\item{applyPropLand}{boolean}

\item{applyPropValue}{boolean}

\item{aggGear}{boolean}

\item{userGears}{data frame}

\item{fleetDescription}{character string}

\item{unkVar}{character vector}

\item{knStrata}{character vector}

\item{disaggSkatesHakes}{boolean}
}
\value{
Data frame (data.table) (n x 12)
Each row of the data.table represents a species record for a given tow/trip

\item{YEAR}{Year of trip/tow}
\item{MONTH}{Month of trip/tow}
\item{NEGEAR}{Fishing gear used on trip/tow}
\item{TONCL2}{Tonnage class of the fishing vessel (2 digit value)}
\item{NESPP3}{Species code (3 charachters)}
\item{MARKET_CODE}{market code (2 characters)}
\item{MESHCAT}{Code to describe the mesh size for a trawl vessel}
\item{AREA/EPU}{Statistical area/ Ecological Production Unit in which species was reportly caught}
\item{UTILCD}{Utilization code}
\item{US}{Landing from the USA vessels or foreign vessels}
\item{SPPLIVLB}{live weight (landed = "n") or landed weight (landed="y") in lbs}
\item{SPPVALUE}{The value of landed catch to the nearest dollar (U.S.), paid to fisherman by dealer, for a given species.}
}
\description{
Connects to cfdbs and pulls fields from WOLANDS, WODETS, CFDETS
}
\section{File Creation}{


A file containing the data.table above will also be saved to the users machine in the directory provided
}

