% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dealers.R
\name{get_dealers}
\alias{get_dealers}
\title{Extract Dealers location details}
\usage{
get_dealers(channel, state = NA, year = NA)
}
\arguments{
\item{channel}{an Object inherited from \code{ROracle::Oracle}. This object is used to connect
to communicate with the database engine. (see \code{dbutils::connect_to_database})}

\item{state}{Character vector. State abbreviations. Default = NA (All states)}

\item{year}{Numeric vector. Years in which to pull data. Defaul = NA (All years)}
}
\value{
A list is returned:

\item{data}{containing the result of the executed \code{sqlStatement}}

\item{sql}{containing the sql call}

\item{colNames}{a vector of the table's column names}

If no \code{sqlStatement} is provided the default sql statement "\code{select * from NEFSC_GARFO.PERMIT_DEALER}" is used
}
\description{
Extract a list of dealer names and locations from supporting tables
}
\section{Reference}{

Use the data dictionary for field name explanations
}

\examples{
\dontrun{
# extracts dealer location table
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_dealers(channel)

# extracts dealer details for states, Massachusetts and Maine
get_dealers(channel,state = c("MA","ME"))

# extracts dealer details for years 2010 to 2020
get_dealers(channel, year = 2010:2020)

# extracts dealer details for years 2010 to 2020 for MA & ME
get_dealers(channel, state = c("MA","ME"), year = 2010:2020)
}

}
\seealso{
\code{\link[dbutils]{connect_to_database}}

Other get functions: 
\code{\link{get_areas}()},
\code{\link{get_gears}()},
\code{\link{get_locations}()},
\code{\link{get_ports}()},
\code{\link{get_species}()},
\code{\link{get_species_itis}()},
\code{\link{get_species_stock_area}()},
\code{\link{get_vessels}()}
}
\concept{get functions}
