% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foreign_data.R
\name{get_foreign_data}
\alias{get_foreign_data}
\title{Downloads all NAFO data}
\usage{
get_foreign_data(
  filterByYear = NA,
  filterByArea = NA,
  removeUSA = T,
  aggregateCountry = T
)
}
\arguments{
\item{filterByYear}{Numeric vector. Years for which data is required}

\item{filterByArea}{Character vector. NAFO Areas for which data is required}

\item{removeUSA}{Boolean. Should USA landings be removed from data set? (Default = T, remove)}

\item{aggregateCountry}{Boolean. Should all catch be aggregated over country codes? (Default = T)}
}
\value{
Data frame: NAFO data

\item{Year}{Year of catch}
\item{MONTH}{Month of catch}
\item{QY}{Quatere year of catch. Jan-Mar = 1, ..., Oct-Dec = 4}
\item{GearCode}{NAFO gear code}
\item{Tonnage}{Size class of vessel}
\item{DivCode}{Division code in which vessel reported catch}
\item{Code}{NAFO species code of landed fish}
\item{SPPLIVMT}{catch in Metric tons}
\item{Country}{Reporting country - only if \code{aggregateCounty = F}}
}
\description{
Downloads, imports, aggregates NAFO data from 21B data base
}
\examples{
nafoData <- get_foreign_data(filterByYear = 2000)
head(nafoData)

}
\seealso{
NAFO 21B website: \url{https://www.nafo.int/Data/Catch-Statistics-STATLANT-21B}
}
