% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_herring_data.R
\name{get_herring_data}
\alias{get_herring_data}
\title{Pull Herring data}
\usage{
get_herring_data(channel, filterByYear = NA, filterByArea = NA)
}
\arguments{
\item{channel}{an Object inherited from \code{ROracle::Oracle}. This object is used to connect
to communicate with the database engine. (see \code{dbutils::connect_to_database})}

\item{filterByYear}{numeric vector. Years to be retrieved (Default = NA, pull all years)}

\item{filterByArea}{numeric vector. Statistical Areas to be retrieved (Default = NA, pull all areas)}
}
\value{
Herring data as a data frame
}
\description{
Herring Data comes from the state of Maine. The underlying oracle table is
updated periodically, often when an assessment is required. The herring data
from CAMS/STOCKEFF is incomplete. This is used in its place
Pulled from NEFSC_GARFO.maine_herring_catch
}
