% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Extract LOCATION information from CFDBS}
\usage{
get_locations(channel)
}
\arguments{
\item{channel}{an Object inherited from \code{ROracle::Oracle}. This object is used to connect
to communicate with the database engine. (see \code{dbutils::connect_to_database})}
}
\value{
A list is returned:

\item{data}{containing the result of the executed \code{sqlStatement} }

\item{sql}{containing the \code{sqlStatement} itself}

\item{colNames}{a vector of the table's column names}
}
\description{
Extract a list of lat, long, ten minute square, etc from the NEFSC "loc" supporting table
}
\section{Reference}{

Use the data dictionary for field name explanations
}

\examples{
\dontrun{
# extracts complete locations table based on default sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_locations(channel)
}

}
\seealso{
\code{\link[dbutils]{connect_to_database}}

Other get functions: 
\code{\link{get_areas}()},
\code{\link{get_gears}()},
\code{\link{get_ports}()},
\code{\link{get_species}()},
\code{\link{get_species_itis}()},
\code{\link{get_species_stock_area}()},
\code{\link{get_vessels}()}
}
\concept{get functions}
