% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mendaden_data.R
\name{get_menhaden_data}
\alias{get_menhaden_data}
\title{Pull Menhaden data from FOSS API}
\usage{
get_menhaden_data(filterByYear = NA, filterByArea = NA)
}
\arguments{
\item{filterByYear}{numeric vector. Years to be retrieved (Default = NA, pull all years)}

\item{filterByArea}{numeric vector. Statistical Areas to be retrieved (Default = NA, pull all areas)}
}
\value{
Menhaden data as a data frame
}
\description{
Pulls Menhaden data from the FOSS API
}
\section{FOSS API}{


You can use the Fisheries One Stop Shop (\href{https://www.fisheries.noaa.gov/foss/f?p=215:200:1011421373172:Mail::::}{FOSS}) database through
a website or an API. You select a time period, a geographic region and a species to get the data you need
}

