% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_stock_area.R
\name{get_species_stock_area}
\alias{get_species_stock_area}
\title{Extract Statistical Areas that comprise species stock definition}
\usage{
get_species_stock_area(channel, species = "all", stock_name = NULL)
}
\arguments{
\item{channel}{an Object inherited from \code{ROracle::Oracle}. This object is used to connect
to communicate with the database engine. (see \code{dbutils::connect_to_database})}

\item{species}{A specific species code or set of codes. Either numeric or character vector. Defaults to "all" species.
Numeric codes (SPECIES_ITIS) are converted to VARCHAR2 (6) when creating the sql statement.}

\item{nameType}{Character string. Upper or lower case. Either "common_name" (default) or "itis".
Determines which type of name to search under.}
}
\value{
A list is returned:

\item{data}{containing the result of the executed \code{sqlStatement}}

\item{sql}{containing the sql call}

\item{colNames}{ a vector of the table's column names}

The default sql statement "\code{select * from STOCKEFF.V_CF_STOCK_RECENT_STAT_AREA_O}" is used
}
\description{
Extract a list of statistical areas from STOCKEFF supporting table
}
\section{Reference}{

Use the data dictionary for field name explanations
}

\examples{
\dontrun{
# extracts complete area table based on default sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_species_stock_area(channel)

# extracts info for cod (164712)
get_species_stock_area(channel,species=164712)

# extracts info for cod ("COD"). All stocks
get_species_stock_area(channel,"cod")
get_species_stock_area(channel,"co")
get_species_stock_area(channel,"COD")

# extracts info for cod (GBK stock)
get_species_stock_area(channel,"COD", stock_name = "GBK")
get_species_stock_area(channel,"CO", stock_name = "gbk")
get_species_stock_area(channel,"164712", stock_name = "EGOM")
get_species_stock_area(channel,164712, stock_name = "egom")

# extracts info for cod (164712)  and bluefish (168559)
sqlStatement <- "select * from cfdbs.species_itis_ne"
get_species_stock_area(channel,species= c("164712","168559"))

}

}
\seealso{
\code{\link[dbutils]{connect_to_database}}

Other get functions: 
\code{\link{get_areas}()},
\code{\link{get_gears}()},
\code{\link{get_locations}()},
\code{\link{get_ports}()},
\code{\link{get_species}()},
\code{\link{get_species_itis}()},
\code{\link{get_vessels}()}
}
\concept{get functions}
