% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessels.R
\name{get_vessels}
\alias{get_vessels}
\title{Extract VESSEL information}
\usage{
get_vessels(channel)
}
\arguments{
\item{channel}{an Object inherited from \code{ROracle::Oracle}. This object is used to connect
to communicate with the database engine. (see \code{dbutils::connect_to_database})}
}
\value{
A list is returned:

\item{data}{containing the result of the executed \code{sqlStatement} }

\item{sql}{containing the \code{sqlStatement} itself}

\item{colNames}{a vector of the table's column names}
}
\description{
Extract a list of vessel ID's, tonnage, crew size, home port, etc from the NEFSC_GARFO "PERMIT_VPS_VESSEL" supporting table
}
\section{Reference}{

Use the data dictionary for field name explanations
}

\examples{
\dontrun{
# extracts complete vessel table based on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_vessels(channel)

}

}
\seealso{
\code{\link[dbutils]{connect_to_database}}

Other get functions: 
\code{\link{get_areas}()},
\code{\link{get_dealers}()},
\code{\link{get_gears}()},
\code{\link{get_locations}()},
\code{\link{get_ports}()},
\code{\link{get_species}()},
\code{\link{get_species_itis}()},
\code{\link{get_species_stock_area}()}
}
\concept{get functions}
