% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskeyAreas.R
\docType{data}
\name{mskeyAreas}
\alias{mskeyAreas}
\title{mskeyAreas:  Area Species, EPU designations}
\format{
A Data frame
\describe{
    \item{AREA}{Statistical Area}
    \item{NESPP3}{Gulf of Maine EPU}
    \item{MeanProp}{Proportion of Statistical area to attribute to EPU}
    \item{EPU}{Ecological Production Unit}
}
}
\usage{
mskeyAreas
}
\description{
Option to pass to \code{get_comland_data} if the user wishes define proportion Statistical Areas
based on findings in mskeyrun project. Statistical Area that straddle EPUs have landings proportioned among
neighboring EPUs
}
\keyword{datasets}
