% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskeyGears.R
\docType{data}
\name{mskeyGears}
\alias{mskeyGears}
\title{mskeyGears:  Gear Fleet mapping from mskeyrun project}
\format{
A Data frame
\describe{
\item{NEGEAR2}{NEFSC 2 digit Gear code}
\item{MESHCAT}{Category of mesh size, SM, LG}
\item{Fleet}{Name of Fleet}
}
}
\usage{
mskeyGears
}
\description{
Associates NEGEAR2 codes to Fleets
}
\details{
and not to proportion the stat areas to EPUs
}
\keyword{datasets}
