% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comland.R
\name{plot_comland}
\alias{plot_comland}
\title{plots of comlands data}
\usage{
plot_comland(data)
}
\arguments{
\item{data}{data frame. Output from \code{comland.r}}
}
\description{
plots of comlands data
}
