% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_foreign_data_skate_hake.R
\name{process_foreign_data_skate_hake}
\alias{process_foreign_data_skate_hake}
\title{Processes NAFO data to fix skates and hakes}
\usage{
process_foreign_data_skate_hake(nafoland)
}
\arguments{
\item{nafoland}{Data frame. output from \code{\url{process_foreign_data}}}
}
\value{
Data frame: NAFO data
}
\description{
Apportions skates and hakes in the same way commercial data is handled - using survey
}
