% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soeAreas.R
\docType{data}
\name{soeAreas}
\alias{soeAreas}
\title{soeAreas:  Area Species,EPU designations}
\format{
A Data frame
\describe{
\item{AREA}{Statistical Area}
\item{NESPP3}{Gulf of Maine EPU}
\item{MeanProp}{Proportion of Statistical area to attribute to EPU}
\item{EPU}{Ecological Production Unit}
}
}
\usage{
soeAreas
}
\description{
Option to pass to \code{get_comland_data} if the user wishes define EPUs exclusively
by Stat areas and not to proportion the landings in these Statistical areas to EPUs
}
\keyword{datasets}
