% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gear.R
\name{aggregate_gear}
\alias{aggregate_gear}
\title{Assign landing records to an aggregated area}
\usage{
aggregate_gear(comland, userGears, fleetDescription)
}
\arguments{
\item{comland}{Data set generated by \code{get_comland_data}}

\item{userAreas}{Data frame. Definitions to aggregate statistical areas to user defined
areas}

\item{areaDescription}{Character. Name of column in userAreas that defines the new
area.}

\item{propDescription}{Character. Name of column in userAreas that defines the 
proportions of landings assigned to new area.}
}
\description{
Takes the output from \code{get_comland_data} and further aggregates from NAFO 
statistical areas to a user defined area.  Allows for species to be assigned by
proportions to more than two user defined areas from one stat area
}
