% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_catch_area.R
\name{assign_catch_area}
\alias{assign_catch_area}
\title{assign unknown catch using known catch characteristics}
\usage{
assign_catch_area(comland.agg)
}
\arguments{
\item{comland}{Data frame. master data frame containing species landings}
}
\value{
updated comland data frame
}
\description{
Uses area. Expand .....
}
\seealso{
Other assign catch: \code{\link{assign_catch_gear}},
  \code{\link{assign_catch_qy_hy}},
  \code{\link{assign_catch_size}}
}
\concept{assign catch}
