% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_catch_gear.R
\name{assign_catch_gear}
\alias{assign_catch_gear}
\title{assign unknown catch using known catch characteristics}
\usage{
assign_catch_gear(comland.agg)
}
\value{
updated comland
}
\description{
Uses gear. Expand .....
}
