% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_catch_qy_hy.R
\name{assign_catch_qy_hy}
\alias{assign_catch_qy_hy}
\title{assign unknown catch using known catch characteristics}
\usage{
assign_catch_qy_hy(comland.agg)
}
\value{
updated comland
}
\description{
Uses Half year and Quarter year. Expand .....
}
