% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_catch_size.R
\name{assign_catch_size}
\alias{assign_catch_size}
\title{assign unknown catch using known catch characteristics}
\usage{
assign_catch_size(comland.agg)
}
\arguments{
\item{comland}{Data frame. master data frame containing species landings}
}
\value{
updated comland data frame
}
\description{
Uses size. Expand .....
}
\seealso{
Other assign catch: \code{\link{assign_catch_area}},
  \code{\link{assign_catch_gear}},
  \code{\link{assign_catch_qy_hy}}
}
\concept{assign catch}
