% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comland.r
\name{comland}
\alias{comland}
\title{Comland.r Version now controlled by git - originally part of comcatch.r}
\usage{
comland(channel, GEARS = GEARs, EPUS = EPUs, use.existing = "y",
  landed = "y", foreign = "y", adjust.ppi = "y", sum.by = "EPU",
  endyear = 2018, reftime = c(2016, 1), out.dir = here::here(),
  Stand.alone = F)
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{GEARS}{List. Designates the NEGEAR codes that comprise a fishing fleet. Default = GEARs (lazily loaded data)}

\item{EPUS}{List. Designates the stat areas that comprise an EPU. Default = EPUs (lazily loaded data)}

\item{use.existing}{String. Pull from database "n" or use existing pull "y" (saves time) . Default = "y"}

\item{landed}{Character String. Use landed weight for scallops and clams instead of live weight. Default = "y"}

\item{foreign}{Character String. Mark foreign landings and keep seperate. Default = "y"}

\item{adjust.ppi}{Character String. Adjust value for inflation. Default = "y"}

\item{sum.by}{Character String. Variable to sum landings by either "EPU" (Default) or "stat.area"}

\item{endyear}{Numeric Scalar. Final year of query. Default = 2018}

\item{reftime}{Numeric Vector. (Length 2). Specifies the year and month if adjusting for inflation. Default = c(2016,1)}

\item{out.dir}{Character string. Path to directory where final output will be saved or where data is to be read from}

\item{Stand.alone}{Boolean. Flag to determine whether to save Skate and hake data to file. defualt = F (Both a US catch file and a NAFO catch file will be saved)}
}
\description{
#Grab commercial landings data from US and Foreign countries (NAFO)
Need to fix menhaden data
SML
}
