% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comland_separate_hakes.R
\name{comland_separate_hakes}
\alias{comland_separate_hakes}
\title{#Comcatch_skates_hakes.r}
\usage{
comland_separate_hakes(comland, skate.hake.us)
}
\description{
Determine proportion of little/winter skates and silver hake in landings data 7/13
SML
}
