% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comlandr.R
\docType{package}
\name{comlandr}
\alias{comlandr}
\title{comlandr: Pull and process commercial fisheries data}
\description{
For use in multispecies projects.
}
\details{
\itemize{
\item pulls data from cfdbs: by area, gear, species, year, month
\item pulls Herring data from state of Maine database
\item defines fleet structure and assigns fleets to gear types
\item define geographic areas and assigns statistical areas to geographic areas
\item assigns catch to unknown gear, area, quarter/half year, size based on similar trips (based on M Palmer, ????)
\item pulls and processes NAFO (Northwest Atlantic Fisheries Organization) data
\item uses survey data to apportion hake complex into species
\item uses survey data to apportion skate complex into species
\item adjusts species value to specified date
\item plots output
}
}
\section{References}{


Palmer, M (????) Working doc reference and title

NAFO website: \url{https://www.nafo.int}

#'To learn more about using \code{comlandr}, start with the vignette: \code{browseVignettes(package="comlandr")} or click the index below
}

