% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foreign_data.R
\name{get_foreign_data}
\alias{get_foreign_data}
\title{Downloads all NAFO data}
\usage{
get_foreign_data(channel)
}
\value{
Data frame: NAFO data

\item{}{}
\item{}{}
\item{}{}
\item{}{}
\item{}{}
}
\description{
Downloads and reads in all NAFO data then aggregates it
}
