% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foreign_data.R
\name{get_foreign_data}
\alias{get_foreign_data}
\title{Downloads all NAFO data}
\usage{
get_foreign_data(removeUSA = T, aggregateCountry = T)
}
\arguments{
\item{removeUSA}{Boolean. Should USA landings be removed from data set (Default = T, remove)}

\item{aggregateCountry}{Boolean. Should all catch be aggregated over country codes? (Default = T)}
}
\value{
Data frame: NAFO data

\item{Year}{Year of catch}
\item{MONTH}{Month of catch}
\item{QY}{Quatere year of catch. Jan-Mar = 1, ..., Oct-Dec = 4}
\item{GearCode}{NAFO gear code}
\item{Tonnage}{Size class of vessel}
\item{DivCode}{Division code in which vessel reported catch}
\item{NAFOCode}{NAFO species code of landed fish}
\item{SPPLIVMT}{catch in Metric tons}
\item{Country}{Reporting country - only if \code{aggregateCounty = F}}
\item{NESPP3}{NEFSC species code}
}
\description{
Downloads, imports, aggregates NAFO data
}
