% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_foreign_data.R,
%   R/process_foreign_data_redo.R
\name{process_foreign_data}
\alias{process_foreign_data}
\title{Processes NAFO data for comlandr use}
\usage{
process_foreign_data(channel, nafoland, EPUs = NULL)

process_foreign_data(channel, nafoland, EPUs = NULL)
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link[dbutils]{connect_to_database}})}

\item{nafoland}{Data frame. output from \code{\url{get_foreign_data}}}

\item{EPUs}{Data frame. Currently a place holder}
}
\value{
Data frame: NAFO data

Data frame: NAFO data
}
\description{
Assigns NAFO Gear codes to NEGEAR codes and NEGEAR codes to fleets
Assigns NAFO Species to NESPP3 species
Assigns NAFO Divcodes to EPUs
Assigns NAFO tonnage to tonnage classes

Adds cfdbs NESPP3, NEGEAR, TONCL1 codes to data.
Adds EPUs
}
\details{
Fills in missing data by including from 21A (SS EPU only)
}
